+++ config/auth-log.php
@@ -14,6 +14,9 @@ return [
         'password',
         //add credentials here to be ommitted from being stored in the log database
     ],
+    'credentialsToEncrypt' => [
+        // 'password'
+    ],
     'fieldsToOmit' => [
         //add fields here to be omitted from being stored in the log database, eg:
         // 'user_ip',