+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Can now log ip addresses of all machines that triggered events
 - Can select fields to be omitted, for example, `user_ip`
 - Can now log User Registered events, Lockout events and Password Reset events
+- Can now specify credentials to be encrypted before being stored in the database
 - Can now specify accepted guards within config file
 - Fixed field omissions not working correctly
 - Added `createWithConfigFilters` method to `AuthenticationLogRecord`